l10n_ao/models/account_tax_regime.py
from odoo import models,fields, api


class TaxRegime(models.Model):

    _name = "account.tax.regime"
    _description = "Account Tax Regime"

    name = fields.Char(string="Name", required=True, translate=True)
