l10n_ao/requirements.txt
xmlschema
dicttoxml2
pycryptodome
